Successful Build State info
There are some important notes about this build state:

First, VirtualPlayers was moved and the build order was changed because
BattleScoreboardAPI depends on VirtualPlayers, therefore, VirtualPlayers
should be compiled first.  Also, an old build of VirtualPlayers (1.6.7)
was added to the binaries to be installed because that's currently, the
specific version of VirtualPlayers that BattleScoreboardAPI requires.

** IMPORTANT **

This build exploit a bug in the maven install plugin version 2.5.2 and
older whereby it would NOT read/write the POM from the JAR file, but
rather, it would just generate a generic POM.

Consenquently, explicit dependencies must be used (as was changed in
WorldEditUtil).  And obviously, changing the version of the maven
install plugin will break the build because Maven is looking for
transitive dependencies that aren't even required for the build.