The real purpose of .gitignore
Files & folders that are commonly ignored between two different projects
should reside in the ~/.gitignore_global in order to reduce duplication.

The real purpose of .gitignore is to ignore files & folders that are
unique to the project:  In this case, BattlePluginsDev auto-generates
folders for the test /server/ and it also automatically installs files
into the /server/plugins/ folder.  Therefore, since the /server/ folder
is unique to BattlePluginsDev (not commonly generated by other types of
projects), we should ignore this folder.

Please put any other files &/or folders that are not specific to
BattlePluginsDev into your .gitignore_global file.